requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
mysql-connector-python==9.0.0
openai==1.35.14
packaging==23.2
pluggy==1.3.0
pstruc==1.0.1
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pytest==7.4.3
python-dotenv==1.0.1
PyYAML==6.0.1
setuptools==71.0.1
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.4
typing_extensions==4.12.2
